+++ packages/lib/src/types/types.ts
@@ -31,6 +31,7 @@ import type { AnyType } from "./schemas"
 import { typesMaybe, typesMaybeNull } from "./utility/maybe"
 import { OrTypeInfo, typesOr } from "./utility/or"
 import { RefinementTypeInfo, typesRefinement } from "./utility/refinement"
+import { TagTypeInfo, typesTag } from "./utility/tag"
 import { typesUnchecked, UncheckedTypeInfo } from "./utility/unchecked"
 export { getTypeInfo } from "./getTypeInfo"
 export { TypeInfo } from "./TypeChecker"
@@ -42,6 +43,7 @@ export {
   StringTypeInfo,
   FrozenTypeInfo,
   ObjectMapTypeInfo,
+  TagTypeInfo,
   RefinementTypeInfo,
   RecordTypeInfo,
   RefTypeInfo,
@@ -73,6 +75,7 @@ export const types = {
   ref: typesRef,
   frozen: typesFrozen,
   enum: typesEnum,
+  tag: typesTag,
   refinement: typesRefinement,
   integer: typesInteger,
   nonEmptyString: typesNonEmptyString,